requirements.txt
streamlit>=1.28.0
llama-index>=0.9.0
llama-index-embeddings-ollama>=0.1.0
llama-index-llms-ollama>=0.1.0
llama-index-vector-stores-chroma>=0.1.0
chromadb>=0.4.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-ollama>=0.1.0
langgraph>=0.0.17
pymupdf>=1.23.0
requests>=2.31.0
pillow>=10.0.0
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0